refactor: Split internal modules
